+++ backend/PapersFeed_backend/__init__.py
@@ -1,3 +1,5 @@
+# pylint: skip-file
+
 from __future__ import absolute_import, unicode_literals
 
 # This will make sure the app is always imported when
