+++ backend/PapersFeed_backend/celery.py
@@ -9,7 +9,7 @@ from celery import Celery
 # set the default Django settings module for the 'celery' program.
 os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'PapersFeed_backend.settings')
 
-app = Celery('papersfeed')
+app = Celery('papersfeed') # pylint: disable=invalid-name
 
 # Using a string here means the worker doesn't have to serialize
 # the configuration object to child processes.
@@ -19,8 +19,3 @@ app.config_from_object(settings, namespace='CELERY')
 
 # Load task modules from all registered Django app configs.
 app.autodiscover_tasks()
-
-
-@app.task(bind=True)
-def debug_task(self):
-    print('Request: {0!r}'.format(self.request))
