+++ backend/PapersFeed_backend/settings.py
@@ -140,7 +140,8 @@ CSRF_COOKIE_NAME = 'csrftoken'
 
 # CELERY
 CELERY_BROKER_URL = 'redis://localhost:6379/0'
-CELERY_BROKER_BACKEND = 'redis://localhost:6379/0'
+# CELERY_BROKER_BACKEND = 'redis://localhost:6379/0'
+CELERY_RESULT_BACKEND = 'redis://localhost:6379/0'
 CELERY_ACCEPT_CONTENT = ['application/json']
 CELERY_TASK_SERIALIZER = 'json'
 CELERY_RESULT_SERIALIZER = 'json'
