+++ frontend/public/index.html
@@ -7,7 +7,7 @@
     <meta name="theme-color" content="#000000" />
     <meta
       name="description"
-      content="PapersFeed"
+      content="Browse and Organize papers, Share and Expand your ideas"
     />
     <link rel="apple-touch-icon" href="logo192.png" />
     <!--
